#3 app.ts모듈분리
각 관심사에 따라 필요한 코드끼리 모듈 분리 및 import설정
전역 객체 window선언을 통한 store변수 공유
