+++ src/pages/index.ts
@@ -1,0 +1,2 @@
+export {default as NewsFeedView}  from './news-feeds-view';
+export {default as NewsDetailView}  from './news-detail-view';
